backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
sentencepiece==0.1.99
numpy==1.24.3
scikit-learn==1.3.2
boto3==1.29.7
python-dotenv==1.0.0
aiofiles==23.2.1
python-json-logger==2.0.7

